re-organizing to project references as opposed to nuget references in canal related projects for unit testing
