+++ Panama.Canal.Tests.Modules/Sagas/CreateFoo/Entries/ReviewCreateFooAnswerEntry.cs
@@ -3,7 +3,8 @@ using Panama.Canal.Extensions;
 using Panama.Canal.Models.Messaging;
 using Panama.Canal.Sagas.Stateless.Extensions;
 using Panama.Canal.Sagas.Stateless.Interfaces;
-using Panama.Canal.Tests.Models;
+using Panama.Canal.Tests.Modules.Models;
+using Panama.Canal.Tests.Modules.Sagas.CreateFoo.Events;
 using Panama.Canal.Tests.Sagas.CreateFoo.Triggers;
 using Panama.Extensions;
 using Panama.Interfaces;
