+++ Panama.Canal.Tests.Modules/Sagas/CreateFoo/Events/ReviewCreateFooAnswerEvent.cs
@@ -3,13 +3,13 @@ using Panama.Canal.Extensions;
 using Panama.Canal.Models.Messaging;
 using Panama.Canal.Sagas.Stateless.Extensions;
 using Panama.Canal.Sagas.Stateless.Interfaces;
-using Panama.Canal.Tests.Models;
+using Panama.Canal.Tests.Modules.Models;
 using Panama.Canal.Tests.Sagas.CreateFoo.States;
 using Panama.Extensions;
 using Panama.Interfaces;
 using Panama.Models;
 
-namespace Panama.Canal.Tests.Sagas.CreateFoo.Events
+namespace Panama.Canal.Tests.Modules.Sagas.CreateFoo.Events
 {
     public class ReviewCreateFooAnswerEvent : ISagaStepEvent
     {
