+++ Panama.Canal.Tests.Modules/Sagas/CreateFoo/Events/RollbackFooEvent.cs
@@ -4,16 +4,15 @@ using Panama.Canal.Extensions;
 using Panama.Canal.Interfaces;
 using Panama.Canal.Sagas.Stateless.Extensions;
 using Panama.Canal.Sagas.Stateless.Interfaces;
-using Panama.Canal.Tests.Models;
+using Panama.Canal.Tests.Modules.Models;
 using Panama.Canal.Tests.Sagas.CreateFoo.States;
-using Panama.Canal.Tests.Sagas.CreateFoo.Triggers;
 using Panama.Extensions;
 using Panama.Interfaces;
 using Panama.Models;
 
-namespace Panama.Canal.Tests.Sagas.CreateFoo.Events
+namespace Panama.Canal.Tests.Modules.Sagas.CreateFoo.Events
 {
-    public class CreateFooEvent : ISagaStepEvent
+    public class RollbackFooEvent : ISagaStepEvent
     {
         public async Task<ISagaState> Execute(IContext context)
         {
@@ -26,20 +25,19 @@ namespace Panama.Canal.Tests.Sagas.CreateFoo.Events
                     .Data(model)
                     .Channel(channel)
                     .Token(context.Token)
-                    .Topic("foo.created")
-                    .Trigger<ReviewCreateFooAnswer>()
-                    .State<CreateFooRequestAnswered>()
-                .Post();
+                    .Topic("foo.failed")
+                    .Post();
+
                 await channel.Commit();
             }
 
             var state = context.Provider.GetService<State>();
             if (state == null)
-                return context.GetState<CreateFooRequested>();
-            
-            state.Data.Add(new Kvp<string, string>("saga.event.name", nameof(CreateFooEvent)));
+                return context.GetState<CreateFooComplete>();
 
-            return context.GetState<CreateFooRequested>();
+            state.Data.Add(new Kvp<string, string>("saga.event.name", nameof(RollbackFooEvent)));
+
+            return context.GetState<CreateFooComplete>();
         }
     }
 }
