+++ Panama.Canal.Tests.Modules/Sagas/CreateFoo/States/CreateFooCreated.cs
