+++ Panama.Canal.Tests.Modules/Sagas/CreateFoo/States/CreateFooFailed.cs
