+++ Panama.Canal.Tests.Modules/Sagas/CreateFoo/States/CreateFooRequested.cs
