+++ Panama.Canal.Tests.Modules/Sagas/CreateFoo/States/CreateFooRollbackRequested.cs
