+++ Panama.Canal.Tests.Modules/Sagas/CreateFoo/Triggers/CompleteNewFoo.cs
