+++ Panama.Canal.Tests.Modules/Sagas/CreateFoo/Triggers/CreateNewFoo.cs
