+++ Panama.Canal.Tests.Modules/Sagas/CreateFoo/Triggers/PublishCreateFooResults.cs
