+++ Panama.Canal.Tests.Modules/Subscriptions/FooFailed.cs
@@ -2,34 +2,33 @@
 using Microsoft.Extensions.Logging;
 using Panama.Canal.Attributes;
 using Panama.Canal.Interfaces;
-using Panama.Canal.Tests.Models;
+using Panama.Canal.Tests.Modules.Models;
 using Panama.Extensions;
 using Panama.Interfaces;
 using Panama.Models;
 
-namespace Panama.Canal.Tests.Subscriptions
+namespace Panama.Canal.Tests.Modules.Subscriptions
 {
-    [DefaultTopic("foo.ack")]
-    public class FooAcknowledged : ISubscribe
+    [DefaultTopic("foo.failed")]
+    public class FooFailed : ISubscribe
     {
         private readonly ILogger<FooCreated> _log;
         private readonly IServiceProvider _provider;
 
-        public FooAcknowledged(
+        public FooFailed(
               IServiceProvider provider
             , ILogger<FooCreated> log)
         {
             _log = log;
             _provider = provider;
         }
-
         public Task Event(IContext context)
         {
-            var kvp = new Kvp<string, string>("subscription.name", nameof(FooAcknowledged));
-            
+            var kvp = new Kvp<string, string>("subscription.name", nameof(FooFailed));
+
             context.Add(kvp);
 
-            _log.LogInformation($"{typeof(FooAcknowledged)} subscriber executed.");
+            _log.LogInformation($"{typeof(FooFailed)} subscriber executed.");
 
             var state = _provider.GetService<State>();
             if (state == null)
