+++ Panama.Canal.Tests.MySQL/MySqlBusTests.cs
@@ -7,11 +7,9 @@ using Panama.Canal.Extensions;
 using Panama.Canal.Interfaces;
 using Panama.Canal.MySQL.Extensions;
 using Panama.Canal.MySQL.Models;
-using Panama.Canal.Tests.Models;
-using Panama.Canal.Tests.Subscriptions;
+using Panama.Canal.Tests.Modules.Models;
+using Panama.Canal.Tests.Modules.Subscriptions;
 using Panama.Extensions;
-using Panama.Models;
-using Polly;
 using System;
 using System.Data;
 using System.Linq;
