+++ Panama.Canal.Tests.MySQL/MySqlMessageTests.cs
@@ -9,8 +9,8 @@ using Panama.Canal.Interfaces;
 using Panama.Canal.Models.Messaging;
 using Panama.Canal.MySQL.Extensions;
 using Panama.Canal.MySQL.Models;
-using Panama.Canal.Tests.Models;
-using Panama.Canal.Tests.Subscriptions;
+using Panama.Canal.Tests.Modules.Models;
+using Panama.Canal.Tests.Modules.Subscriptions;
 using Panama.Extensions;
 using System;
 using System.Collections.Generic;
