+++ Panama.Canal.Tests.MySQL/Panama.Canal.Tests.MySQL.csproj
@@ -40,14 +40,14 @@
     <PackageReference Include="NLog.Database" Version="5.1.4" />
     <PackageReference Include="NLog.Extensions.Logging" Version="5.2.3" />
     <PackageReference Include="NLog.Web.AspNetCore" Version="5.2.3" />
-    <PackageReference Include="Panama.Canal" Version="6.0.4" />
-    <PackageReference Include="Panama.Canal.MySQL" Version="6.0.3" />
     <PackageReference Include="Pomelo.EntityFrameworkCore.MySql" Version="6.0.2" />
     <PackageReference Include="System.Configuration.ConfigurationManager" Version="7.0.0" />
    </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\Panama.Canal.Tests\Panama.Canal.Tests.csproj" />
+    <ProjectReference Include="..\Panama.Canal.MySQL\Panama.Canal.MySQL.csproj" />
+    <ProjectReference Include="..\Panama.Canal.Tests.Modules\Panama.Canal.Tests.Modules.csproj" />
+    <ProjectReference Include="..\Panama.Canal\Panama.Canal.csproj" />
   </ItemGroup>
 
 </Project>
