+++ Panama.Canal.Tests.RabbitMQ/Subscriptions/BarCreated.cs
@@ -1,7 +1,7 @@
 ﻿using Microsoft.Extensions.Logging;
 using Panama.Canal.Interfaces;
 using Panama.Canal.RabbitMQ.Attributes;
-using Panama.Canal.Tests.Models;
+using Panama.Canal.Tests.Modules.Models;
 using Panama.Extensions;
 using Panama.Interfaces;
 using Panama.Models;
