+++ Panama.Canal.Tests/MessageTests.cs
@@ -9,8 +9,8 @@ using Panama.Canal.Jobs;
 using Panama.Canal.Models;
 using Panama.Canal.Models.Messaging;
 using Panama.Canal.Models.Options;
-using Panama.Canal.Tests.Models;
-using Panama.Canal.Tests.Subscriptions;
+using Panama.Canal.Tests.Modules.Models;
+using Panama.Canal.Tests.Modules.Subscriptions;
 using Panama.Extensions;
 using Panama.Models;
