+++ Panama.Canal.Tests/SagaTests.cs
@@ -6,11 +6,11 @@ using Panama.Canal.Extensions;
 using Panama.Canal.Interfaces;
 using Panama.Canal.Sagas.Extensions;
 using Panama.Canal.Sagas.Stateless.Extensions;
-using Panama.Canal.Tests.Models;
-using Panama.Canal.Tests.Sagas.CreateFoo;
-using Panama.Canal.Tests.Sagas.CreateFoo.Events;
-using Panama.Canal.Tests.Sagas.CreateFoo.Exits;
-using Panama.Canal.Tests.Subscriptions;
+using Panama.Canal.Tests.Modules.Models;
+using Panama.Canal.Tests.Modules.Sagas.CreateFoo;
+using Panama.Canal.Tests.Modules.Sagas.CreateFoo.Events;
+using Panama.Canal.Tests.Modules.Sagas.CreateFoo.Exits;
+using Panama.Canal.Tests.Modules.Subscriptions;
 using Panama.Extensions;
 using Panama.Models;
